Ignore [binding] routing keys
This exchange never advertised (or probably intended) to use routing keys
for, well, anything. But currently it does and it gets confusing. Since
the intent is to pick a random binding's queue, perhaps it makes
more sense to load all bindings, the same way the fanout exchange
implementation does.

Without it, binding routing keys have an unexpected effect on the outcome.

Per discussion with @gerhard.